--- a/Cql/CodeGeneration.NET/CSharpSourceCodeWriter.cs
+++ b/Cql/CodeGeneration.NET/CSharpSourceCodeWriter.cs
@@ -200,35 +200,12 @@ namespace Hl7.Cql.CodeGeneration.NET
             if (PartialClass)
                 writer.WriteLine(indentLevel, $"partial class {className}");
             else
-                writer.WriteLine(indentLevel, $"public class {className}");
+                writer.WriteLine(indentLevel, $"public static class {className}");
             writer.WriteLine(indentLevel, "{");
             writer.WriteLine();
             indentLevel += 1;
             // Class
             {
-                writer.WriteLine();
-
-                writer.WriteLine(indentLevel, $"{AccessModifierString(ContextAccessModifier)} CqlContext context;");
-                writer.WriteLine();
-                writeCachedValues(definitions, libraryName, writer, indentLevel);
-
-                // Write constructor
-                writer.WriteLine(indentLevel, $"public {className}(CqlContext context)");
-                writer.WriteLine(indentLevel, "{");
-                {
-                    indentLevel += 1;
-
-                    writer.WriteLine(indentLevel, "this.context = context ?? throw new ArgumentNullException(\"context\");");
-                    writer.WriteLine();
-
-                    writeDependencies(dependencyGraph, libraryNameToClassName, libraryName, writer, indentLevel);
-                    writer.WriteLine();
-                    writeCachedValueNames(definitions, libraryName, writer, indentLevel);
-                    indentLevel -= 1;
-                }
-                writer.WriteLine(indentLevel, "}");
-
-                WriteLibraryMembers(writer, dependencyGraph, libraryName, libraryNameToClassName!, indentLevel);
                 writeMemoizedInstanceMethods(definitions, libraryName, writer, indentLevel);
                 indentLevel -= 1;
                 writer.WriteLine(indentLevel, "}");
@@ -247,62 +224,7 @@ namespace Hl7.Cql.CodeGeneration.NET
                 }
             }
         }
-
-        private void writeCachedValueNames(DefinitionDictionary<LambdaExpression> definitions, string libraryName, StreamWriter writer, int indentLevel)
-        {
-            foreach (var kvp in definitions.DefinitionsForLibrary(libraryName))
-            {
-                foreach (var overload in kvp.Value)
-                {
-                    if (isDefinition(overload.Item2))
-                    {
-                        var methodName = VariableNameGenerator.NormalizeIdentifier(kvp.Key);
-                        var cachedValueName = DefinitionCacheKeyForMethod(methodName!);
-                        var returnType = ExpressionConverter.PrettyTypeName(overload.Item2.ReturnType);
-                        var privateMethodName = PrivateMethodNameFor(methodName!);
-                        writer.WriteLine(indentLevel, $"{cachedValueName} = new Lazy<{returnType}>(this.{privateMethodName});");
-                    }
-                }
-            }
-        }
-
-        private static void writeDependencies(DirectedGraph dependencyGraph, Func<string?, string?> libraryNameToClassName, string libraryName, StreamWriter writer, int indentLevel)
-        {
-            var node = dependencyGraph.Nodes[libraryName];
-            var requiredLibraries = node.ForwardEdges?
-                .Select(edge => edge.ToId)
-                .Except(new[] { dependencyGraph.EndNode.NodeId })
-                .Distinct();
-            foreach (var dependentLibrary in requiredLibraries!)
-            {
-                var typeName = libraryNameToClassName!(dependentLibrary);
-                var memberName = typeName;
-                writer.WriteLine(indentLevel, $"{memberName} = new {typeName}(context);");
-            }
-        }
-
-        private void writeCachedValues(DefinitionDictionary<LambdaExpression> definitions, string libraryName, StreamWriter writer, int indentLevel)
-        {
-            writer.WriteLine(indentLevel, "#region Cached values");
-            writer.WriteLine();
-            var accessModifier = AccessModifierString(DefinesAccessModifier);
-            foreach (var kvp in definitions.DefinitionsForLibrary(libraryName))
-            {
-                foreach (var overload in kvp.Value)
-                {
-                    if (isDefinition(overload.T))
-                    {
-                        var methodName = VariableNameGenerator.NormalizeIdentifier(kvp.Key);
-                        var cachedValueName = DefinitionCacheKeyForMethod(methodName!);
-                        var returnType = ExpressionConverter.PrettyTypeName(overload.T.ReturnType);
-                        writer.WriteLine(indentLevel, $"{accessModifier} Lazy<{returnType}> {cachedValueName};");
-                    }
-                }
-            }
-            writer.WriteLine();
-            writer.WriteLine(indentLevel, "#endregion");
-        }
-
+        
         private void writeTupleTypes(IEnumerable<Type> tupleTypes, Func<string, Stream> libraryNameToStream, bool closeStream)
         {
             if (tupleTypes.Any())
@@ -346,36 +268,6 @@ namespace Hl7.Cql.CodeGeneration.NET
             overload.Parameters.Count == 1
                 && overload.Parameters[0].Type == typeof(CqlContext);
 
-        private void WriteLibraryMembers(TextWriter writer,
-            DirectedGraph dependencyGraph,
-            string libraryName,
-            Func<string, string> libraryNameToClassName,
-            int indent)
-        {
-            var node = dependencyGraph.Nodes[libraryName];
-            var requiredLibraries = node.ForwardEdges?
-                .Select(edge => edge.ToId)
-                .Except(new[] { dependencyGraph.EndNode.NodeId })
-                .Distinct();
-            if (requiredLibraries != null)
-            {
-
-                writer.WriteLine(indent, "#region Dependencies");
-                writer.WriteLine();
-
-                foreach (var dependentLibrary in requiredLibraries)
-                {
-                    var typeName = libraryNameToClassName(dependentLibrary);
-                    var memberName = typeName;
-                    writer.WriteLine(indent, $"public {typeName} {memberName} {{ get; }}");
-                }
-
-                writer.WriteLine();
-                writer.WriteLine(indent, "#endregion");
-                writer.WriteLine();
-            }
-        }
-
         private IList<DirectedGraphNode> DetermineBuildOrder(DirectedGraph minimalGraph)
         {
             var sorted = minimalGraph.TopologicalSort()
@@ -384,14 +276,6 @@ namespace Hl7.Cql.CodeGeneration.NET
                 .ToList();
             return sorted;
         }
-
-        private string DefinitionCacheKeyForMethod(string methodName)
-        {
-            if (methodName[0] == '@')
-                return "__" + methodName.Substring(1);
-            else return "__" + methodName;
-        }
-        private string PrivateMethodNameFor(string methodName) => methodName + "_Value";
 
         private void WriteMemoizedInstanceMethod(string libraryName, TextWriter writer, int indentLevel,
             string cqlName,
@@ -419,16 +303,6 @@ namespace Hl7.Cql.CodeGeneration.NET
 
             if (isDef)
             {
-                // Definitions, which are CQL expressions without parameter, can be memoized,
-                // so we generate a "generator" function (name ending in _Value) and a
-                // getter function, which just calls triggers the lazy to invoke this
-                // first _Value method.
-                var cachedValueName = DefinitionCacheKeyForMethod(methodName!);
-                var privateMethodName = PrivateMethodNameFor(methodName!);
-
-                var func = expressionConverter.ConvertTopLevelFunctionDefinition(indentLevel, overload, privateMethodName, "private");
-                writer.Write(func);
-                writer.WriteLine();
                 writer.WriteLine(indentLevel, $"[CqlDeclaration(\"{cqlName}\")]");
                 WriteTags(writer, indentLevel, tags);
 
@@ -447,20 +321,14 @@ namespace Hl7.Cql.CodeGeneration.NET
                     }
                 }
 
-                var lazyType = typeof(Lazy<>).MakeGenericType(visitedBody.Type);
-                var valueFunc =
-                    Expression.Lambda(
-                    Expression.MakeMemberAccess(
-                    Expression.Parameter(lazyType, cachedValueName),
-                    lazyType.GetMember("Value").Single()));
-
-                writer.Write(expressionConverter.ConvertTopLevelFunctionDefinition(indentLevel, valueFunc, methodName!, "public"));
+                var func = expressionConverter.ConvertTopLevelFunctionDefinition(indentLevel, overload, methodName!, "public static");
+                writer.Write(func);
             }
             else
             {
                 writer.WriteLine(indentLevel, $"[CqlDeclaration(\"{cqlName}\")]");
                 WriteTags(writer, indentLevel, tags);
-                writer.Write(expressionConverter.ConvertTopLevelFunctionDefinition(indentLevel, overload, methodName!, "public"));
+                writer.Write(expressionConverter.ConvertTopLevelFunctionDefinition(indentLevel, overload, methodName!, "public static"));
                 //      writer.WriteLine();
             }
         }