… .idea/.idea.Cql/.idea/.gitignore
… .idea/.idea.Cql/.idea/.name
… .idea/.idea.Cql/.idea/indexLayout.xml
… .idea/.idea.Cql/.idea/vcs.xml
… Cql/CoreTests/CoreTests.csproj
… Cql/Cql.Compiler/Cql.Compiler.csproj
… Cql/Cql.Firely/FhirCqlContext.cs
… Cql/Cql.Firely/FhirModelBindingSetup.cs
… Cql/Cql.Firely/FhirTypeConverter.cs
… Cql/Cql.Logging/Cql.Logging.csproj
… Cql/Cql.Packaging/LibraryPackager.cs
… Cql/Graph/Graph.csproj
… Cql/PackagerCLI/PackagerCLI.csproj
… Demo/Measures/AdultOutpatientEncountersFHIR4-2.2.000.cs
… Demo/Measures/AdvancedIllnessandFrailtyExclusionECQMFHIR4-5.17.000.cs
… Demo/Measures/BCSEHEDISMY2022-1.0.0.cs
… Demo/Measures/BreastCancerScreeningsFHIR-0.0.009.cs
… Demo/Measures/CervicalCancerScreeningFHIR-0.0.005.cs
… Demo/Measures/ColorectalCancerScreeningsFHIR-0.0.003.cs
… Demo/Measures/CumulativeMedicationDurationFHIR4-1.0.000.cs
… Demo/Measures/DRCommunicationWithPhysicianManagingDiabetesFHIR-0.0.004.cs
… Demo/Measures/DevDays-2023.0.0.cs
… Demo/Measures/DiabetesHemoglobinA1cHbA1cPoorControl9FHIR-0.0.015.cs
… Demo/Measures/DischargedonAntithromboticTherapyFHIR-0.0.010.cs
… Demo/Measures/Exam125FHIR-0.0.009.cs
… Demo/Measures/Exam130FHIR-0.0.003.cs
… Demo/Measures/FHIR347-0.1.021.cs
… Demo/Measures/FHIRHelpers-4.0.001.cs
… Demo/Measures/HospiceFHIR4-2.3.000.cs
… Demo/Measures/HospitalHarmHyperglycemiainHospitalizedPatientsFHIR-0.0.006.cs